# Some micro optimizations in db.go and node.go