From the collapseTransactions benchmarks, I notice some slice allocations that can be reduced and the cost of defer() in node.go.

By improving the slice allocations and remove the use defer() in node.rehashNoWrite(), I saw there are some performance improvements, albeit minor in some cases.

Following are the statistics of the benchmarks. 
The benchmarks are ran 100 times.

<br />

Remove defer() in node.rehashNoWrite()
```
name             old time/op    new time/op    delta
RehashNoWrite-8     151ns ± 6%     120ns ± 5%  -20.39%  (p=0.000 n=96+95)
```

<br />

Preallocate the slices

<br />

```
name                 old time/op    new time/op    delta
ReadUnderAccounts-8    77.4ns ± 7%    42.4ns ± 8%  -45.24%  (p=0.000 n=86+98)

name                 old alloc/op   new alloc/op   delta
ReadUnderAccounts-8     96.0B ± 0%     48.0B ± 0%  -50.00%  (p=0.000 n=100+100)

name                 old allocs/op  new allocs/op  delta
ReadUnderAccounts-8      2.00 ± 0%      1.00 ± 0%  -50.00%  (p=0.000 n=100+100)
```

<br />

```
name                  old time/op    new time/op    delta
WriteUnderAccounts-8     327ns ± 0%     320ns ±26%   -2.12%  (p=0.000 n=64+97)

name                  old alloc/op   new alloc/op   delta
WriteUnderAccounts-8      240B ± 0%      192B ± 0%  -20.00%  (p=0.000 n=100+100)

name                  old allocs/op  new allocs/op  delta
WriteUnderAccounts-8      3.00 ± 0%      2.00 ± 0%  -33.33%  (p=0.000 n=100+100)
```

<br />

```
name                       old time/op    new time/op    delta
ReadAccountContractPage-8     217ns ±16%     198ns ± 8%   -8.96%  (p=0.000 n=88+86)

name                       old alloc/op   new alloc/op   delta
ReadAccountContractPage-8      624B ± 0%      608B ± 0%   -2.56%  (p=0.000 n=100+100)

name                       old allocs/op  new allocs/op  delta
ReadAccountContractPage-8      4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.000 n=100+100)
```

<br />

```
name                        old time/op    new time/op    delta
WriteAccountContractPage-8     731ns ± 3%     699ns ± 2%   -4.44%  (p=0.000 n=86+97)

name                        old alloc/op   new alloc/op   delta
WriteAccountContractPage-8      848B ± 0%      832B ± 0%   -1.89%  (p=0.000 n=100+100)

name                        old allocs/op  new allocs/op  delta
WriteAccountContractPage-8      4.00 ± 0%      3.00 ± 0%  -25.00%  (p=0.000 n=100+100)
```