@iwasaki-kenta 2019-10-03
Looks perfectly reasonable as some minor improvements. `defer` has been sped up over recent revision in Go, though for the time being we can merge this PR in for the tests and reduction in allocs in other areas.